Added authenticated activity. Removed unused functionality on sign in.
